fix: rename entryList component to overviewEntryList
